requirements.txt
fastapi
uvicorn
six
pyyaml
python-dotenv
docker